Added Point text func

+++ point_text.go
@@ -3,8 +3,8 @@ package sprint
 import "fmt"
 
 type Point struct {
-	X    float32
-	Y    float32
+	X    float64
+	Y    float64
 	Text string
 }
 
@@ -12,6 +12,6 @@ func PointText(p Point) Point {
 	return Point{
 		X:    p.X,
 		Y:    p.Y,
-		Text: fmt.Sprintf("Text at (%.7f, %.7f)", p.X, p.Y),
+		Text: fmt.Sprintf("Text at (%.6f, %.6f)", p.X, p.Y),
 	}
 }
